additional_code/requirements.txt
biopython==1.83
contourpy==1.1.1
csubst @ git+https://github.com/kfuku52/csubst@740bd4cf346d5fd71bf28cff6c47b537186ace08
cycler==0.12.1
Cython==3.0.8
ete3==3.1.3
fonttools==4.47.2
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
Pyvolve==1.1.0
scipy==1.10.1
seaborn==0.13.1
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.4
zipp==3.17.0
